+++ protocols/gossipsub/src/topic.rs
@@ -24,7 +24,7 @@ use prost::Message;
 use sha2::{Digest, Sha256};
 use std::fmt;
 
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
+#[derive(Debug, Clone, PartialEq, Eq, Hash, PartialOrd, Ord)]
 pub struct TopicHash {
     /// The topic hash. Stored as a string to align with the protobuf API.
     hash: String,
@@ -35,17 +35,13 @@ impl TopicHash {
         TopicHash { hash: hash.into() }
     }
 
-    pub fn into_string(self) -> String {
-        self.hash
-    }
-
     pub fn as_str(&self) -> &str {
         &self.hash
     }
 }
 
 /// A gossipsub topic.
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord)]
 pub struct Topic {
     topic: String,
 }
@@ -80,14 +76,20 @@ impl Topic {
     }
 }
 
+impl Into<String> for TopicHash {
+    fn into(self) -> String {
+        self.hash
+    }
+}
+
 impl fmt::Display for Topic {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         write!(f, "{}", self.topic)
     }
 }
 
 impl fmt::Display for TopicHash {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         write!(f, "{}", self.hash)
     }
 }