+++ protocols/kad/CHANGELOG.md
@@ -1,5 +1,7 @@
 # 0.22.0 [unreleased]
 
+- Update `libp2p-core` and `libp2p-swarm` dependencies.
+
 - Add `KBucketRef::range` exposing the minimum inclusive and maximum inclusive
   `Distance` for the bucket
   ([PR 1680](https://github.com/libp2p/rust-libp2p/pull/1680)).