+++ transports/wasm-ext/CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.21.0 [unreleased]
+
+- Update `libp2p-core` dependency.
+
 # 0.20.1 [2020-07-06]
 
 - Improve the code quality of the `websockets.js` binding with the browser's `WebSocket` API.